Moving helper functions out of bin script for dashboard
