+++ lib/chainsync/chainsync/dashboard/__init__.py
@@ -1,5 +1,8 @@
 """Dashboard utilities"""
 
+from .build_leaderboard import build_leaderboard
+from .build_ticker import build_ticker
 from .extract_data_logs import calculate_spot_price, get_combined_data, read_json_to_pd
 from .plot_fixed_rate import plot_fixed_rate
 from .plot_ohlcv import plot_ohlcv
+from .usernames import address_to_username, combine_usernames, get_user_lookup
